+++ src/main/java/insane96mcp/insanelib/InsaneLib.java
@@ -6,6 +6,7 @@ import insane96mcp.insanelib.setup.ClientSetup;
 import insane96mcp.insanelib.setup.Config;
 import insane96mcp.insanelib.setup.ILEntities;
 import insane96mcp.insanelib.setup.ILGlobalLootModifiers;
+import net.minecraft.client.Minecraft;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
@@ -13,10 +14,14 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
 
 @Mod(InsaneLib.MOD_ID)
 public class InsaneLib
@@ -25,9 +30,12 @@ public class InsaneLib
     public static final String MOD_ID = "insanelib";
     public static final String RESOURCE_PREFIX = MOD_ID + ":";
 
+    public static DecimalFormat ONE_DECIMAL_FORMATTER;
+
     public InsaneLib() {
         ModLoadingContext.get().registerConfig(net.minecraftforge.fml.config.ModConfig.Type.COMMON, Config.COMMON_SPEC);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(ClientSetup::init);
+        FMLJavaModLoadingContext.get().getModEventBus().addListener(InsaneLib::clientSetup);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::preInit);
         MinecraftForge.EVENT_BUS.register(this);
         final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
@@ -43,5 +51,11 @@ public class InsaneLib
     public void onAddReloadListener(AddReloadListenerEvent event) {
         JsonFeatureDataReloadListener.reloadContext = event.getConditionContext();
         event.addListener(JsonFeatureDataReloadListener.INSTANCE);
+    }
+
+    @SubscribeEvent
+    public static void clientSetup(FMLClientSetupEvent event) {
+        DecimalFormatSymbols DECIMAL_FORMAT_SYMBOLS = new DecimalFormatSymbols(Minecraft.getInstance().getLocale());
+        ONE_DECIMAL_FORMATTER = new DecimalFormat("#.#", DECIMAL_FORMAT_SYMBOLS);
     }
 }